+++ test/index.test.js
@@ -7,6 +7,7 @@ const start = require('./common');
 const assert = require('assert');
 const random = require('./util').random;
 const stream = require('stream');
+const { EventEmitter } = require('events');
 
 const collection = 'blogposts_' + random();
 
@@ -1018,6 +1019,71 @@ describe('mongoose module:', function() {
         const optionsSentToMongo = nativeAggregateSpy.args[0][1];
         assert.equal(optionsSentToMongo.allowDiskUse, false);
       });
+    });
+    describe('global `timestamps.createdAt.immutable` (gh-10139)', () => {
+      it('is `true` by default', () => {
+        // Arrange
+        const m = new mongoose.Mongoose();
+
+        // Act
+        const userSchema = new m.Schema({ name: String }, { timestamps: true });
+
+        // Assert
+        assert.equal(userSchema.path('createdAt').options.immutable, true);
+      });
+
+      it('can be overridden to `false`', () => {
+        // Arrange
+        const m = new mongoose.Mongoose();
+        m.set('timestamps.createdAt.immutable', false);
+
+        // Act
+        const userSchema = new m.Schema({ name: String }, { timestamps: true });
+
+        // Assert
+        assert.equal(userSchema.path('createdAt').options.immutable, false);
+      });
+    });
+  });
+
+  describe('custom drivers', function() {
+    it('can set custom driver (gh-11900)', async function() {
+      const m = new mongoose.Mongoose();
+
+      class Collection {
+        findOne(filter, options, cb) {
+          cb(null, { answer: 42 });
+        }
+      }
+      class Connection extends EventEmitter {
+        constructor(base) {
+          super();
+          this.base = base;
+          this.models = {};
+        }
+
+        collection() {
+          return new Collection();
+        }
+
+        openUri(uri, opts, callback) {
+          this.readyState = mongoose.ConnectionStates.connected;
+          callback();
+        }
+      }
+      const driver = {
+        Collection,
+        getConnection: () => Connection
+      };
+
+      m.setDriver(driver);
+
+      await m.connect();
+
+      const Test = m.model('Test', m.Schema({ answer: Number }));
+
+      const res = await Test.findOne();
+      assert.deepEqual(res.toObject(), { answer: 42 });
     });
   });
 });