+++ test/query.middleware.test.js
@@ -636,7 +636,6 @@ describe('query middleware', function() {
     });
     const Model = db.model('Test', schema);
 
-
     await Model.find();
     assert.equal(called, 1);
 
@@ -654,5 +653,47 @@ describe('query middleware', function() {
 
     await Model.aggregate([{ $match: { name: 'test' } }]);
     assert.equal(called, 3);
+  });
+
+  it('allows skipping the wrapped function with `skipMiddlewareFunction()` (gh-11426)', async function() {
+    const schema = Schema({ name: String });
+    const now = Date.now();
+
+    schema.pre('find', function(next) {
+      next(mongoose.skipMiddlewareFunction([{ name: 'from cache' }]));
+    });
+    schema.post('find', function(res) {
+      res.forEach(doc => {
+        doc.loadedAt = now;
+      });
+    });
+    const Test = db.model('Test', schema);
+
+    const res = await Test.find();
+    assert.equal(res.length, 1);
+    assert.strictEqual(res[0].name, 'from cache');
+    assert.strictEqual(res[0].loadedAt, now);
+  });
+
+  it('allows overwriting result with `overwriteMiddlewareResult()` (gh-11426)', async function() {
+    const schema = Schema({ name: String });
+
+    schema.post('updateOne', function() {
+      return mongoose.overwriteMiddlewareResult({ answer: 42 });
+    });
+    schema.post('updateOne', function(res) {
+      assert.strictEqual(res.answer, 42);
+      res.secondMiddlewareRan = true;
+    });
+    const Test = db.model('Test', schema);
+
+    const { _id } = await Test.create({ name: 'test' });
+    const res = await Test.updateOne({ _id }, { name: 'changed' });
+    assert.equal(res.answer, 42);
+    assert.strictEqual(res.modifiedCount, undefined);
+    assert.strictEqual(res.secondMiddlewareRan, true);
+
+    const doc = await Test.findById(_id);
+    assert.equal(doc.name, 'changed');
   });
 });