+++ test/types/schema.test.ts
@@ -1,5 +1,17 @@
-import { Schema, Document, SchemaDefinition, SchemaDefinitionProperty, SchemaTypeOptions, Model, Types } from 'mongoose';
-import { expectType, expectError } from 'tsd';
+import {
+  Schema,
+  Document,
+  SchemaDefinition,
+  SchemaDefinitionProperty,
+  SchemaTypeOptions,
+  Model,
+  Types,
+  InferSchemaType,
+  SchemaType,
+  Query,
+  HydratedDocument
+} from 'mongoose';
+import { expectType, expectError, expectAssignable } from 'tsd';
 
 enum Genre {
   Action,
@@ -320,6 +332,199 @@ function gh10900(): void {
   });
 }
 
+export function autoTypedSchema() {
+  // Test auto schema type obtaining with all possible path types.
+
+  class Int8 extends SchemaType {
+    constructor(key, options) {
+      super(key, options, 'Int8');
+    }
+    cast(val) {
+      let _val = Number(val);
+      if (isNaN(_val)) {
+        throw new Error('Int8: ' + val + ' is not a number');
+      }
+      _val = Math.round(_val);
+      if (_val < -0x80 || _val > 0x7F) {
+        throw new Error('Int8: ' + val +
+          ' is outside of the range of valid 8-bit ints');
+      }
+      return _val;
+    }
+  }
+
+  type TestSchemaType = {
+    string1?: string;
+    string2?: string;
+    string3?: string;
+    string4?: string;
+    number1?: number;
+    number2?: number;
+    number3?: number;
+    number4?: number;
+    date1?: Date;
+    date2?: Date;
+    date3?: Date;
+    date4?: Date;
+    buffer1?: Buffer;
+    buffer2?: Buffer;
+    buffer3?: Buffer;
+    buffer4?: Buffer;
+    boolean1?: boolean;
+    boolean2?: boolean;
+    boolean3?: boolean;
+    boolean4?: boolean;
+    mixed1?: any;
+    mixed2?: any;
+    mixed3?: any;
+    objectId1?: Schema.Types.ObjectId;
+    objectId2?: Schema.Types.ObjectId;
+    objectId3?: Schema.Types.ObjectId;
+    customSchema?: Int8;
+    map1?: Map<string, string>;
+    map2?: Map<string, number>;
+    array1?: string[];
+    array2?: any[];
+    array3?: any[];
+    array4?: any[];
+    array5?: any[];
+    decimal1?: Schema.Types.Decimal128;
+    decimal2?: Schema.Types.Decimal128;
+    decimal3?: Schema.Types.Decimal128;
+  };
+
+  const TestSchema = new Schema({
+    string1: String,
+    string2: 'String',
+    string3: 'string',
+    string4: Schema.Types.String,
+    number1: Number,
+    number2: 'Number',
+    number3: 'number',
+    number4: Schema.Types.Number,
+    date1: Date,
+    date2: 'Date',
+    date3: 'date',
+    date4: Schema.Types.Date,
+    buffer1: Buffer,
+    buffer2: 'Buffer',
+    buffer3: 'buffer',
+    buffer4: Schema.Types.Buffer,
+    boolean1: Boolean,
+    boolean2: 'Boolean',
+    boolean3: 'boolean',
+    boolean4: Schema.Types.Boolean,
+    mixed1: Object,
+    mixed2: {},
+    mixed3: Schema.Types.Mixed,
+    objectId1: Schema.Types.ObjectId,
+    objectId2: 'ObjectId',
+    objectId3: 'objectId',
+    customSchema: Int8,
+    map1: { type: Map, of: String },
+    map2: { type: Map, of: Number },
+    array1: [String],
+    array2: Array,
+    array3: [Schema.Types.Mixed],
+    array4: [{}],
+    array5: [],
+    decimal1: Schema.Types.Decimal128,
+    decimal2: 'Decimal128',
+    decimal3: 'decimal128'
+  });
+
+  type InferredTestSchemaType = InferSchemaType<typeof TestSchema>;
+
+  expectType<TestSchemaType>({} as InferredTestSchemaType);
+
+  const SchemaWithCustomTypeKey = new Schema({
+    name: {
+      customTypeKey: String,
+      required: true
+    }
+  }, {
+    typeKey: 'customTypeKey'
+  });
+
+  expectType<string>({} as InferSchemaType<typeof SchemaWithCustomTypeKey>['name']);
+
+  const AutoTypedSchema = new Schema({
+    userName: {
+      type: String,
+      required: [true, 'userName is required']
+    },
+    description: String,
+    nested: new Schema({
+      age: {
+        type: Number,
+        required: true
+      },
+      hobby: {
+        type: String,
+        required: false
+      }
+    }),
+    favoritDrink: {
+      type: String,
+      enum: ['Coffee', 'Tea']
+    },
+    favoritColorMode: {
+      type: String,
+      enum: {
+        values: ['dark', 'light'],
+        message: '{VALUE} is not supported'
+      },
+      required: true
+    }
+  }, {
+    statics: {
+      staticFn() {
+        expectType<Model<AutoTypedSchemaType['schema']>>(this);
+        return 'Returned from staticFn' as const;
+      }
+    },
+    methods: {
+      instanceFn() {
+        expectType<HydratedDocument<AutoTypedSchemaType['schema']>>(this);
+        return 'Returned from DocumentInstanceFn' as const;
+      }
+    },
+    query: {
+      byUserName(userName) {
+        expectAssignable<Query<unknown, AutoTypedSchemaType['schema']>>(this);
+        return this.where({ userName });
+      }
+    }
+  });
+
+  type InferredSchemaType = InferSchemaType<typeof AutoTypedSchema>;
+
+  expectType<AutoTypedSchemaType['schema']>({} as InferredSchemaType);
+
+  expectError<AutoTypedSchemaType['schema'] & { doesNotExist: boolean; }>({} as InferredSchemaType);
+
+  return AutoTypedSchema;
+}
+
+export type AutoTypedSchemaType = {
+  schema: {
+    userName: string;
+    description?: string;
+    nested?: {
+      age: number;
+      hobby?: string
+    },
+    favoritDrink?: 'Tea' | 'Coffee',
+    favoritColorMode: 'dark' | 'light'
+  }
+  , statics: {
+    staticFn: () => 'Returned from staticFn'
+  },
+  methods: {
+    instanceFn: () => 'Returned from DocumentInstanceFn'
+  },
+};
+
 // discriminator
 const eventSchema = new Schema<{ message: string }>({ message: String }, { discriminatorKey: 'kind' });
 const batchSchema = new Schema<{ name: string }>({ name: String }, { discriminatorKey: 'kind' });
@@ -353,5 +558,4 @@ function gh11828() {
       }
     }
   });
-
 }